requirements.txt
asgiref==3.7.2
attrs==23.2.0
cachetools==5.3.2
Django~=4.2.6
django-admin-extra-buttons==1.5.7
django-cleanup==8.1.0
django-object-actions==4.2.0
exceptiongroup==1.2.0
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-language==2.13.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
httplib2==0.22.0
iniconfig==2.0.0
packaging==23.2
phonenumbers==8.13.29
Pillow==10.2.0
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytest==8.0.0
pytest-django==4.8.0
sqlparse==0.4.4
tzdata==2023.4
whitenoise==6.6.0
django-simple-history==3.4.0
cryptography==42.0.2
websocket-client==1.7.0
requests~=2.31.0
mysql-connector-python~=8.3.0
python-dotenv~=1.0.1
djangorestframework~=3.14.0
django-mysql==4.12.0
